googlecloudstorage_blueprints/requirements.txt
google-cloud-storage==1.28.1
httplib2==0.15.0
